Add per-domain rate limits.
With domain rate limits in place, I think it's safe to test this
toy crawler on somebody else's domains. I hardcoded 3 other small,
static blogs into the seed list, for a total of 4 sites. It crawls
these small sites in about 7 minutes, but most of that time is idle,
waiting for the domain rate limits. I haven't tried a more stressful
crawl, yet.

There is a small bug: when the crawler finishes, it doesn't
correctly shut down all of the coroutines. I need to look into this,
but it doesn't affect crawling itself.