Make the frontier a persistent priority queue.
The frontier is stored in RethinkDB. Although frontier data is
saved to disk, I need to do a bit more work to resume a job after
the crawler restarts.

This step required a lot of re-designing in order to add priorities
and persistence, including fixing the awkward interaction between
the frontier, rate limiter, and job. The new design is more of a
pipeline: job -> frontier -> rate limiter -> downloader.