Allow crawls to be paused, cancelled, and resumed.
This required quite a bit of refactoring to handle cancellation
correctly. In the process, I completely refactored the watchdog
process and the main entrypoint for the application. I also had
to refactor the task hierarchy and the mechanism for returning
items from the downloader to the crawl controller. I fixed a few
outstanding bugs along the way and also added support for the
SIGTERM as well as SIGINT... the former is used by Docker and
is needed for being able to upgrade Starbelly in the field.