This commit adds three new monitoring tools:
  1. CPU Profiler
  2. Resource Monitor
  3. Task Monitor

All features are exposed through the API and supported by the
CLI and GUI clients. I had planned to add these later, but they
turn out to be helpful for debugging/troubleshooting at this
point in development, so they are implemented ahead of schedule.

There are also a few fixes for little bugs I noticed while
working.