Fix "None" exception in downloader
During some production crawls, one of the exceptions that gets logged
is "None", which doesn't make any sense. I added some code to handle
this condition and log additional details so I can get to the bottom of
it. I also added "SocksError" to the list of "short" exceptions -- i.e.
exceptions where we don't need to save a full stack trace.