Solved "None" exception from previous commit
It turns out that the "None" exception is an aiohttp
ServerDisconnectedError, which means the server dropped the TCP
connection before sending the complete response body. The
improved error logging from the previous commit helped
clarify what was going on. I'm removing the None-specific
code from the previous commit and also cleanin up the
error reporting for the login code.