Migrate scheduler to Trio.
Quite a bit of refactoring here and moved some code out of scheduler
and into server or crawl manager, so those will need to be updated.
Starting to lean on trio channels for inter-task communication as
opposed to calling methods. This is a much stricter and cleaner
separation of responsibilities.

Also regenerated protobufs.