Modify rate limiter to use channels.
As part of the port to Trio, I have recognized that I should be using channels
more often as a way to decouple components. The downloader already uses channels
so in this commit I have gone back and modified the rate limiter to use
channels as well. Eventually a crawl job will be able to spin up a few tasks
(frontier, downloader, & extractor) and wire them to each other and the rate
limiter just by hooking up the appropriate channels.