Migrate robots.txt manager to Trio
This is a major milestone because this is the first component that
actually connects to RethinkDB. I took a detour to quickly add Trio
support to RethinkDB (currently in my own private fork) and added
an integration test to ensure that robots.txt manager can talk to
the database.

Also removed robots_txt_manager from policy.py, which simplifies
policy a lot and moves the decision making into robots.py.