Fix timeout bug in downloader
The timeout handling raised an uncaught exception that crashed the
crawler. To simplify things, I've move the timeout handling into the
asyncio part of the downloader, where aiohttp handles it for us, and
gotten rid of the trio timeout code.

Also a fixed a related bug: only successful responses should be
extracted.